+++ Providers/EF6/Revo.EF6/Projections/EF6ProjectionEventListener.cs
@@ -1,23 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Revo.Core.Commands;
-using Revo.Core.Core;
+﻿using Revo.Core.Commands;
 using Revo.Core.Events;
 using Revo.Core.Transactions;
-using Revo.Domain.Entities;
 using Revo.Domain.Events;
 using Revo.Infrastructure.Events.Async;
 using Revo.Infrastructure.Projections;
+using Revo.Infrastructure.Tenancy;
+using System;
+using System.Collections.Generic;
 
 namespace Revo.EF6.Projections
 {
     public class EF6ProjectionEventListener : ProjectionEventListener
     {
-        public EF6ProjectionEventListener(IEF6ProjectionSubSystem projectionSubSystem,
-            IUnitOfWorkFactory unitOfWorkFactory, CommandContextStack commandContextStack,
-            EF6ProjectionEventSequencer eventSequencer) :
-            base(projectionSubSystem, unitOfWorkFactory, commandContextStack)
+        public EF6ProjectionEventListener(Func<IEF6ProjectionSubSystem> projectionSubSystemFunc,
+            IUnitOfWorkFactory unitOfWorkFactory, Func<CommandContextStack> commandContextStackFunc,
+            ITenantProvider tenantProvider, EF6ProjectionEventSequencer eventSequencer) :
+            base(projectionSubSystemFunc, unitOfWorkFactory, commandContextStackFunc, tenantProvider)
         {
             EventSequencer = eventSequencer;
         }