+++ Providers/RavenDB/Revo.RavenDB/Projections/RavenProjectionEventListener.cs
@@ -1,23 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Revo.Core.Commands;
-using Revo.Core.Core;
+﻿using Revo.Core.Commands;
 using Revo.Core.Events;
 using Revo.Core.Transactions;
-using Revo.Domain.Entities;
 using Revo.Domain.Events;
 using Revo.Infrastructure.Events.Async;
 using Revo.Infrastructure.Projections;
+using Revo.Infrastructure.Tenancy;
+using System;
+using System.Collections.Generic;
 
 namespace Revo.RavenDB.Projections
 {
     public class RavenProjectionEventListener : ProjectionEventListener
     {
-        public RavenProjectionEventListener(IRavenProjectionSubSystem projectionSubSystem,
-            IUnitOfWorkFactory unitOfWorkFactory, CommandContextStack commandContextStack,
-            RavenProjectionEventSequencer eventSequencer) :
-            base(projectionSubSystem, unitOfWorkFactory, commandContextStack)
+        public RavenProjectionEventListener(Func<IRavenProjectionSubSystem> projectionSubSystemFunc,
+            IUnitOfWorkFactory unitOfWorkFactory, Func<CommandContextStack> commandContextStackFunc,
+            ITenantProvider tenantProvider, RavenProjectionEventSequencer eventSequencer) :
+            base(projectionSubSystemFunc, unitOfWorkFactory, commandContextStackFunc, tenantProvider)
         {
             EventSequencer = eventSequencer;
         }